Restructure test folder and files
