+++ test/v2/fixture/entries/createResponseData.js
@@ -1,0 +1,15 @@
+const dataSuccessful = {
+  data: {
+    id: 99547865,
+    date: '2021-11-08',
+    hours: 2,
+    notes: 'Issue',
+    task_id: 14521973,
+    user_id: 377264,
+    url: 'https://secure.tickspot.com/114217/api/v2/entries.json',
+    created_at: '2021-11-09T11:30:40.000-05:00',
+    updated_at: '2021-11-09T11:30:40.000-05:00',
+  },
+};
+
+export default dataSuccessful;
