+++ test/v2/fixture/entries/getEntryResponseData.js
@@ -1,17 +1,3 @@
-export const dataSuccessful = {
-  data: {
-    id: 99547865,
-    date: '2021-11-08',
-    hours: 2,
-    notes: 'Issue',
-    task_id: 14521973,
-    user_id: 377264,
-    url: 'https://secure.tickspot.com/114217/api/v2/entries.json',
-    created_at: '2021-11-09T11:30:40.000-05:00',
-    updated_at: '2021-11-09T11:30:40.000-05:00',
-  },
-};
-
 export const dataEntrySuccessful = {
   data: {
     id: 123456,
@@ -45,4 +31,4 @@ export const dataEntryNotFound = {
   },
 };
 
-export default { dataSuccessful, dataEntrySuccessful, dataEntryNotFound };
+export default { dataEntrySuccessful, dataEntryNotFound };
